// backend/src/infrastructure/database/kysely/db.types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface QuotationLines {
  discount_percent: Generated<Numeric>;
  gst_id: string | null;
  id: Generated<string>;
  line_total: Generated<Numeric>;
  product_id: string;
  product_unit_id: string;
  quantity: Numeric;
  quotation_id: string;
  rate: Numeric;
}

export interface Quotations {
  created_at: Generated<Timestamp>;
  created_by: string | null;
  customer_id: string;
  id: Generated<string>;
  quotation_date: Generated<Timestamp>;
  quotation_number: string;
  remarks: string | null;
  status: Generated<string>;
  total_amount: Generated<Numeric>;
  updated_at: Generated<Timestamp>;
  valid_until: Timestamp | null;
}